// Types/DirectURL.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface DirectURL {
	data: Datum[];
}

interface Datum {
	media: Media[];
}

interface Media {
	cipher: Cipher;
	exp: number;
	format: string;
	media_type: string;
	nbf: number;
	sources: Source[];
}

interface Cipher {
	type: string;
}

interface Source {
	provider: string;
	url: string;
}