// Types/SongMetadata.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface SongMetadata {
	error: any[];
	results: Results;
}

export interface Results {
	data: Datum[];
	count: number;
	total: number;
	filtered_count: number;
}

export interface Datum {
	SNG_ID: string;
	ALB_ID: string;
	ALB_PICTURE: string;
	ALB_TITLE: string;
	ARTISTS: Artist[];
	ART_ID: string;
	ART_NAME: string;
	DIGITAL_RELEASE_DATE: string;
	DISK_NUMBER: string;
	DURATION: string;
	EXPLICIT_LYRICS: string;
	EXPLICIT_TRACK_CONTENT: ExplicitTrackContent;
	GENRE_ID: string;
	ISRC: string;
	LYRICS_ID: number;
	PHYSICAL_RELEASE_DATE: string;
	PROVIDER_ID: string;
	RANK_SNG: string;
	SMARTRADIO: number;
	SNG_TITLE: string;
	STATUS: number;
	TRACK_NUMBER: string;
	TYPE: number;
	UPLOAD_ID: number;
	USER_ID: number;
	VERSION: string;
	MD5_ORIGIN: string;
	FILESIZE_AAC_64: string;
	FILESIZE_MP3_64: string;
	FILESIZE_MP3_128: string;
	FILESIZE_MP3_256: string;
	FILESIZE_MP3_320: string;
	FILESIZE_FLAC: string;
	FILESIZE: string;
	GAIN: string;
	MEDIA_VERSION: string;
	TRACK_TOKEN: string;
	TRACK_TOKEN_EXPIRE: number;
	MEDIA: Media[];
	RIGHTS: Rights;
	__TYPE__: string;
}

export interface Artist {
	ART_ID: string;
	ROLE_ID: string;
	ARTISTS_SONGS_ORDER: string;
	ART_NAME: string;
	ARTIST_IS_DUMMY: boolean;
	ART_PICTURE: string;
	RANK: string;
	LOCALES: Locales;
	SMARTRADIO: number;
	__TYPE__: string;
}

export interface Locales {
	"lang_ja-kana": Lang;
	"lang_zh-hant": Lang;
	lang_en: Lang;
	"lang_ja-hrkt": Lang;
	"lang_ja-jpan": Lang;
	lang_ar: Lang;
}

export interface Lang {
	name: string;
}

export interface ExplicitTrackContent {
	EXPLICIT_LYRICS_STATUS: number;
	EXPLICIT_COVER_STATUS: number;
}

export interface Media {
	TYPE: string;
	HREF: string;
}

export interface Rights {
	STREAM_ADS_AVAILABLE: boolean;
	STREAM_ADS: string;
	STREAM_SUB_AVAILABLE: boolean;
	STREAM_SUB: string;
}